template: Parse structured tags when fetching nodes
Add the custom `node.structuredTags` metadata at the very beginning,
when fetching nodes. Instead of calculating the structured tags all
the time. This should be a performance improvement.

https://phabricator.endlessm.com/T31949
