+++ packages/template-ui/src/constants.js
@@ -12,5 +12,7 @@ export const StructuredTags = {
   TYPE: 'type',
 };
 
+export const StructuredTagsRegExp = new RegExp('(.*)=(.*)');
+
 export const SEARCH_MAX_RESULTS = 64;
 export const CAROUSEL_INTERVAL = 10000;
